Rename `ResponseCodes` to `ExpectedResponseCodes`
Add `ExpectedResponseCodes` and `ExpectedDeleteResponseCodes` inputs,
which are more self-explanatory, and mark `ResponseCodes` and
`DeleteResponseCodes` as deprecated in field annotations. Also add a
`Check()` implementation that ensures only one of the inputs is set, and
sets the new field appropriately so we can use the new field
consistently elsewhere in the code.